# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vault API",
        "description": "API Multi-tenant para gestão de contas e transações financeiras",
        "contact": {
            "name": "Vault API Team",
            "email": "dev@vault.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/account/balance": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Consultar saldo",
                "description": "Retorna o saldo atual da conta do usuário autenticado",
                "operationId": "233ac0eca73daead5b725da45b369bfa",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/TenantHeader"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Saldo retornado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "balance": {
                                                    "type": "string",
                                                    "example": "1000.00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login do usuário",
                "description": "Autentica o usuário e retorna um token Bearer",
                "operationId": "dd51a484b7fa0f97c7680fdbde14cf5f",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/TenantHeader"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@alpha.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 6,
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login bem-sucedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|abcdef123456..."
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "João"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "joao@alpha.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas ou tenant não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados de validação inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout do usuário",
                "description": "Revoga o token atual do usuário",
                "operationId": "df9bdc6c7a6d9f20b983526b2ec0f698",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/TenantHeader"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logout bem-sucedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Listar transações",
                "description": "Retorna o histórico de transações do usuário autenticado",
                "operationId": "9b50e29e364179ebde24ac5fc59cd442",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/TenantHeader"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Número máximo de transações a retornar",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de transações",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "transactions": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "type": {
                                                                "type": "string",
                                                                "example": "credit"
                                                            },
                                                            "amount": {
                                                                "type": "string",
                                                                "example": "100.50"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "balance": {
                                                    "type": "string",
                                                    "example": "1000.00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Criar transação",
                "description": "Cria uma nova transação de crédito ou débito na conta do usuário",
                "operationId": "54eaa3952808f639d46f969f4c8b98c5",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/TenantHeader"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "amount"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "credit",
                                            "debit"
                                        ],
                                        "example": "credit"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0.01,
                                        "example": 100.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transação criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transaction created successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "transaction": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "type": {
                                                            "type": "string",
                                                            "example": "credit"
                                                        },
                                                        "amount": {
                                                            "type": "string",
                                                            "example": "100.50"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "balance": {
                                                    "type": "string",
                                                    "example": "1100.50"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados inválidos ou saldo insuficiente para débito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Error": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Error message"
                    }
                },
                "type": "object"
            }
        },
        "parameters": {
            "TenantHeader": {
                "name": "X-Tenant-Slug",
                "in": "header",
                "description": "Slug do tenant (operadora)",
                "required": true,
                "schema": {
                    "type": "string",
                    "example": "alpha"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "sanctum"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Autenticação de usuários"
        },
        {
            "name": "Account",
            "description": "Operações de conta e saldo"
        },
        {
            "name": "Transactions",
            "description": "Operações de crédito, débito e histórico"
        }
    ]
}